models/Molmo/requirements.txt
torch
torchvision
pillow
einops
transformers==4.43.3
